add takestamp sorting for albums

+++ database/migrations/2018_08_15_102039_move_albums.php
@@ -14,19 +14,21 @@ class MoveAlbums extends Migration
      */
     public function up()
     {
-        $results = DB::table('lychee_albums')->select('*')->get();
-        foreach ($results as $result)
-        {
-            $id = $result->id;
-            $id = substr($id,1,10);
-            $id[0] = strval(intval($id[0]) % 4);
-            $album = new Album();
-            $album->id = $id;
-            $album->title = $result->title;
-            $album->description = $result->description;
-            $album->public = $result->public;
-            $album->visible_hidden = $result->visible;
-            $album->save();
+        if(Schema::hasTable('lychee_albums')){
+            $results = DB::table('lychee_albums')->select('*')->get();
+            foreach ($results as $result)
+            {
+                $id = $result->id;
+                $id = substr($id,1,10);
+                $id[0] = strval(intval($id[0]) % 4);
+                $album = new Album();
+                $album->id = $id;
+                $album->title = $result->title;
+                $album->description = $result->description;
+                $album->public = $result->public;
+                $album->visible_hidden = $result->visible;
+                $album->save();
+            }
         }
     }
 
@@ -37,6 +39,8 @@ class MoveAlbums extends Migration
      */
     public function down()
     {
-        DB::table('albums')->delete();
+        if(Schema::hasTable('lychee_albums')) {
+            DB::table('albums')->delete();
+        }
     }
 }
